Fixes 8 bit image mode setting in datasets
Squashed commit of the following:

commit 5524d88c1094254a60c507ceb2d60b1d2429ddbb
Author: Benjamin Kiessling <mittagessen@l.unchti.me>
Date:   Mon Apr 8 20:09:28 2024 +0200

    Shared im_mode setting in GroundTruthDataset/PolygonGTDataset

commit 46be12cabf553a54c9544c4dbaa4078b159d73b7
Author: Benjamin Kiessling <mittagessen@l.unchti.me>
Date:   Sun Apr 7 23:33:45 2024 +0200

    put image mode in shared memory

+++ kraken/lib/util.py
@@ -56,7 +56,7 @@ def is_bitonal(im: Union[Image.Image, torch.Tensor]) -> bool:
     if isinstance(im, Image.Image):
         return im.getcolors(2) is not None and len(im.getcolors(2)) == 2
     elif isinstance(im, torch.Tensor):
-        return len(im.int().unique()) == 2
+        return len(im.unique()) == 2
 
 
 def get_im_str(im: Image.Image) -> str:
